feat: generated key pairs on installation
