+++ src/components/Installbutton.tsx
@@ -1,5 +1,5 @@
 // src/components/InstallButton.tsx
-import React from 'react';
+import React from "react";
 
 interface InstallButtonProps {
   deferredPrompt: Event | null;
@@ -12,10 +12,10 @@ const InstallButton: React.FC<InstallButtonProps> = ({ deferredPrompt }) => {
       deferredPrompt.prompt(); // Show the install prompt
       // @ts-expect-error error
       deferredPrompt.userChoice.then((choiceResult: { outcome: string }) => {
-        if (choiceResult.outcome === 'accepted') {
-          console.log('User accepted the A2HS prompt');
+        if (choiceResult.outcome === "accepted") {
+          console.log("User accepted the A2HS prompt");
         } else {
-          console.log('User dismissed the A2HS prompt');
+          console.log("User dismissed the A2HS prompt");
         }
       });
     }
@@ -24,7 +24,10 @@ const InstallButton: React.FC<InstallButtonProps> = ({ deferredPrompt }) => {
   return (
     <div>
       {deferredPrompt && (
-        <button onClick={handleInstallClick} className="bg-green-500 hover:bg-green-700 text-white font-bold py-2 px-4 rounded">
+        <button
+          onClick={handleInstallClick}
+          className="bg-green-500 hover:bg-green-700 text-white font-bold py-2 px-4 rounded"
+        >
           Install App
         </button>
       )}
